pyproject.toml
[tool.poetry]
name = "template"
version = "0.1.0"
description = ""
authors = [

]
maintainers = [

]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.100.0"
uvicorn = { version = "^0.22.0", extras = ["standard"] }
pydantic = {extras = ["email"], version = "^2.5.3"}
pydantic-settings = "^2"
yarl = "^1.9.2"
ujson = "^5.8.0"
SQLAlchemy = {version = "^2.0.18", extras = ["asyncio"]}
alembic = "^1.11.1"
aiomysql = "^0.2.0"
mysqlclient = "^2.2.0"
httptools = "^0.6.0"
loguru = "^0.7.0"
taskiq = "^0"
taskiq-fastapi = "^0"
pyzmq = "^25"
sqlalchemy-utils = "^0.41.1"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-dotenv = "^1.0.1"
jinja2 = "^3.1.3"


[tool.poetry.dev-dependencies]
pytest = "^7.2.1"
flake8 = "~4.0.1"
mypy = "^1.1.1"
isort = "^5.11.4"
pre-commit = "^3.0.1"
wemake-python-styleguide = "^0.17.0"
black = "^22.12.0"
autoflake = "^1.6.1"
pytest-cov = "^4.0.0"
anyio = "^3.6.2"
pytest-env = "^0.8.1"
httpx = "^0.23.3"
taskiq = { version = "^0", extras = ["reload"] }

[tool.isort]
profile = "black"
multi_line_output = 3
src_paths = ["template",]

[tool.mypy]
strict = true
ignore_missing_imports = true
allow_subclassing_any = true
allow_untyped_calls = true
pretty = true
show_error_codes = true
implicit_reexport = true
allow_untyped_decorators = true
warn_unused_ignores = false
warn_return_any = false
namespace_packages = true

[tool.pytest.ini_options]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
    "ignore:.*unclosed.*:ResourceWarning",
]
env = [
    "TEMPLATE_ENVIRONMENT=pytest",
    "TEMPLATE_DB_BASE=template_test",
]

[fastapi-template.options]
project_name = "template"
api_type = "rest"
enable_redis = "None"
enable_rmq = "None"
ci_type = "none"
enable_migrations = "True"
enable_taskiq = "True"
enable_kube = "None"
kube_name = "template"
enable_routers = "True"
enable_kafka = "None"
enable_loguru = "True"
traefik_labels = "None"
add_dummy = "None"
orm = "sqlalchemy"
self_hosted_swagger = "None"
prometheus_enabled = "None"
sentry_enabled = "None"
otlp_enabled = "None"
pydanticv1 = "None"
gunicorn = "None"
add_users = "None"
cookie_auth = "None"
jwt_auth = "None"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
